+++ src/chevah_compat/posix_filesystem.py
@@ -16,6 +16,7 @@ import struct
 import sys
 import time
 import unicodedata
+import six
 from os import scandir
 
 from zope.interface import implementer
@@ -133,7 +134,8 @@ class PosixFilesystemBase(object):
         '''See `ILocalFilesystem`.'''
 
         if not self._avatar:
-            return self._pathSplitRecursive(str(os.path.expanduser('~')))
+            return self._pathSplitRecursive(
+                six.text_type(os.path.expanduser('~')))
 
         if self._avatar.root_folder_path is None:
             return self._pathSplitRecursive(self._avatar.home_folder_path)
@@ -175,7 +177,7 @@ class PosixFilesystemBase(object):
         if path is None or path == '' or path == '.':
             return self.home_segments
 
-        if not isinstance(path, str):
+        if not isinstance(path, six.text_type):
             path = path.decode(self.INTERNAL_ENCODING)
 
         if not path.startswith('/'):
@@ -230,7 +232,8 @@ class PosixFilesystemBase(object):
             # Check for the virtual segments, but also for any ancestor.
             while target_segments:
                 inside_path = os.path.join(self._root_path, *target_segments)
-                if not os.path.lexists(self.getEncodedPath(inside_path)):
+                encoded_path = self.getEncodedPath(inside_path)
+                if not os.path.lexists(encoded_path):
                     target_segments.pop()
                     continue
                 virtual_path = '/' + '/'.join(virtual_segments)
@@ -344,23 +347,29 @@ class PosixFilesystemBase(object):
         raise NotImplementedError('You must implement this method.')
 
     def getAbsoluteRealPath(self, path):
-        '''See `ILocalFilesystem`.'''
+        """
+        See `ILocalFilesystem`.
+        """
         absolute_path = os.path.abspath(self.getEncodedPath(path))
-
-        if not isinstance(absolute_path, str):
+        if not isinstance(absolute_path, six.text_type):
             absolute_path = absolute_path.decode(self.INTERNAL_ENCODING)
 
         return absolute_path
 
     def isAbsolutePath(self, path):
-        '''See `ILocalFilesystem`.'''
+        """
+        See `ILocalFilesystem`.
+        """
         return os.path.isabs(path)
 
     def isFolder(self, segments):
-        '''See `ILocalFilesystem`.'''
+        """
+        See `ILocalFilesystem`.
+        """
         try:
             return self.getAttributes(segments).is_folder
         except OSError:
+            # On any error, we consider it not a folder.
             return False
 
     def isFile(self, segments):
@@ -742,6 +751,10 @@ class PosixFilesystemBase(object):
             # On Windows, scandir gets float precision while
             # getAttributes only integer.
             modified = int(modified)
+            # On Windows, path might have long names for local drives.
+            # For compat, we keep the simple format as the end user format.
+            if path.startswith('\\\\?\\') and path[5] == ':':
+                path = path[4:]
 
         return FileAttributes(
             name=name,
@@ -766,7 +779,7 @@ class PosixFilesystemBase(object):
         """
         # This is done to allow lazy initialization of process_capabilities.
         from chevah_compat import process_capabilities
-        if not isinstance(name, str):
+        if not isinstance(name, six.text_type):
             name = name.decode(self.INTERNAL_ENCODING)
 
         # OSX HFS+ store file as Unicode, but in normalized format.